split up studio specs
