+++ packages/app/cypress/e2e/studio/helper.ts
@@ -55,6 +55,31 @@ export function launchStudio ({ specName = 'spec.cy.js', createNewTestFromSuite 
   }
 }
 
+export function inputNewTestName (name: string = 'new-test') {
+  cy.findByTestId('new-test-button').click()
+  cy.findByTestId('test-name-input').type(name)
+  cy.findByTestId('create-test-button').click()
+
+  // verify recording is enabled to ensure the panel is fully ready
+  cy.findByTestId('record-button-recording').should('have.text', 'Recording...')
+
+  cy.get('.studio-single-test-container').should('be.visible')
+}
+
+export function incrementCounter (initialCount: number) {
+  cy.getAutIframe().within(() => {
+    cy.get('p').contains(`Count is ${initialCount}`)
+
+    // (1) First Studio action - get
+    cy.get('#increment')
+
+    // (2) Second Studio action - click
+    .realClick().then(() => {
+      cy.get('p').contains(`Count is ${initialCount + 1}`)
+    })
+  })
+}
+
 export function assertClosingPanelWithoutChanges () {
   // Cypress re-runs after you cancel Studio.
   // Original spec should pass
